* develop: (31 commits)
  handle copy and assignment of PiPoStreamAttributes correctly: duplicate label pointer array
  check NULL label
  maxpipohost: fixed get Enum attribute (broken introducing vasrSize Enum)
  copyPiPoAttributes: added check if attr is not NULL
  pipo: added two flags (isArrau and isVarSize) in pipo attr constructors
  maxpipohost: fixed copyPiPoAttributes function about PiPoVarSizeAttr and PiPoArrayAttr
  fix clone method for PiPoVarSizeAttr<const char *>
  specialisation of PiPoVarSizeAttr template for c-strings, replace some c-casts by static_cast
  avoid 2 warnings
  need to include <algorithm> for right definition of std::remove on linux
  avoid implicit conversion warnings long->int by casts or using long
  removed some warnings
  indentation
  conditional compile of C++11 features
  add missing include guard to PiPoParallel.h --> can add pipo.descr to collection
  attr type Dictionary like String
  attr type Dictionary like String
  selectively assign union members, DUH!
  change from getmetric() to converged() method
  PiPoDictionaryAttr: pass initval, never return NULL json
  ...

+++ src/PiPoSequence.h
@@ -117,7 +117,7 @@ public:
     {
       next->setReceiver(receiver);
       
-      for (int i = seq_.size() - 2; i >= 0; i--)
+      for (long i = seq_.size() - 2; i >= 0; i--)
       {
         PiPo *pipo = seq_[i];
         pipo->setReceiver(next);
@@ -135,7 +135,7 @@ public:
   /** @name PiPoChain query methods */
   /** @{ */
 
-  int getSize() const
+  size_t getSize() const
   {
     return seq_.size();
   }
